Renamed alloc_function to unique_function, removed old version, lvalue ref ctor fix
